requirements.txt
fastapi==0.96.0
keras==2.12.0
numpy==1.23.5
Pillow==9.5.0
protobuf==4.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.15
tensorflow==2.12.0
uvicorn
pymysql
google-cloud-storage
python-multipart
mysql-connector-python
gunicorn